null check

+++ portlets/migration-portlet/docroot/WEB-INF/src/de/uhh/l2g/plugins/migration/mapper/FacilityMapper.java
@@ -26,9 +26,10 @@ public class FacilityMapper {
 		if (legacyFacility.getParentId() == 0){
 			try {
 				Institution prevRoot = InstitutionLocalServiceUtil.getRootByGroupId(companyId, groupId);
-				InstitutionLocalServiceUtil.deleteInstitution(prevRoot);
+				if(prevRoot != null){ InstitutionLocalServiceUtil.deleteInstitution(prevRoot);
 				//also clean Link
 				Institution_HostLocalServiceUtil.deleteLinkByInstitution(prevRoot, groupId, companyId);
+				}
 			
 			} catch (Exception e) {
 		         System.out.println(e.getClass().getName());
